+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ unreleased
 ==========
 * feat: Expiry field removed from version table and replaced with compliance number. Added additional content settings action to version table
 * fix: Changed model title to uppercase
+* fix: Compliance number should not be editable when version is not draft
 
 1.3.1 (2022-06-22)
 ==================